pyproject.toml
[tool.poetry]
name = "crewmateservices"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.14.0"
discord = "^1.0.1"
flask = "^1.1.2"
PyNaCl = "^1.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
